// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const QuestionType = {
  SINGLE: 'SINGLE',
  MULTI: 'MULTI',
  SCALE_0_10: 'SCALE_0_10',
  TEXT: 'TEXT',
  MATRIX_AREA: 'MATRIX_AREA'
} as const

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType]


export const ResponseStatus = {
  DRAFT: 'DRAFT',
  COMPLETED: 'COMPLETED'
} as const

export type ResponseStatus = (typeof ResponseStatus)[keyof typeof ResponseStatus]


export const OptionSource = {
  STATIC: 'STATIC',
  AREAS: 'AREAS',
  PROCESOS: 'PROCESOS'
} as const

export type OptionSource = (typeof OptionSource)[keyof typeof OptionSource]